delete no used file
