+++ packages/base/src/page/SyncDataSource/SyncTaskForm/SyncTaskForm.tsx
@@ -22,7 +22,7 @@ import {
   FormItemNoLabel,
   FormItemSubTitle
 } from '@actiontech/shared/lib/components/FormCom';
-import CronInputCom from '@actiontech/shared/lib/components/CronInput/CronInputCom';
+import CronInputCom from '@actiontech/shared/lib/components/CronInput';
 import { Alert, FormInstance, Popconfirm, Spin } from 'antd5';
 import useRuleTemplate from 'sqle/src/hooks/useRuleTemplate';
 import useAuditRequired from '../../../hooks/useAuditRequired';
