--- packages/shared/lib/components/CronInput/index.less
@@ -1,29 +1,0 @@
-.cron-input-wrapper {
-  .cron-inline-select {
-    min-width: 80px;
-  }
-
-  .cron-every-select {
-    width: 100px;
-  }
-}
-
-.cron-week-wrapper {
-  padding-left: 40px;
-}
-
-.cron-inline-select-dropdown {
-  max-width: 300px;
-
-  .rc-virtual-list-holder-inner {
-    flex-flow: row wrap !important;
-  }
-
-  .cron-inline-select-option {
-    flex: 0 0 30px;
-  }
-
-  .anticon-check {
-    display: none;
-  }
-}
